Add some checks and special dirty hacks to make lxde-settings and lxappearance work together.

+++ src/main.c
@@ -11,12 +11,31 @@
 #include <string.h>
 #include <unistd.h>
 
+#include <X11/X.h>
+#include <X11/Xatom.h>
+#include <X11/Xlib.h>
+#include <gdk/gdkx.h>
+
 #include "main-dlg-ui.h"
 #include "demo.h"
 #include "glade-support.h"
 
 char tmp_rc_file[] = "/tmp/gtkrc-2.0-XXXXXX";
 GtkWidget* main_dlg = NULL;
+
+gboolean under_lxde = FALSE;	/* wether lxde-xsettings daemon is active */
+
+static void check_lxde()
+{
+	Atom atom = XInternAtom( GDK_DISPLAY(), "LXDE_SETTINGS", True );
+	if( atom != None )
+	{
+		XGrabServer( GDK_DISPLAY() );
+		if( XGetSelectionOwner( GDK_DISPLAY(), atom ) )
+			under_lxde = TRUE;
+		XUngrabServer( GDK_DISPLAY() );
+	}
+}
 
 int main (int argc, char *argv[])
 {
@@ -33,15 +52,24 @@ int main (int argc, char *argv[])
 
         gtk_set_locale ();
         gtk_init (&argc, &argv);
+
+        check_lxde();
+        if( under_lxde )
+        	return 1;
+
         show_demo( (GdkNativeWindow)atoi( argv[2] ) );
         gtk_main();
         return 0;
     }
 
-    mkstemp( tmp_rc_file );
-
     gtk_set_locale ();
     gtk_init (&argc, &argv);
+
+	/* Dirty hack: check if lxde-xsettings daemon is active */
+	check_lxde();
+
+	if( ! under_lxde )
+		mkstemp( tmp_rc_file );
 
     /* Dirty hack: "gtk-toolbar-style" is installed in class_init of GtkToolbar */
     gtk_widget_destroy( gtk_toolbar_new() );
@@ -52,7 +80,9 @@ int main (int argc, char *argv[])
 
     gtk_main ();
 
-    unlink( tmp_rc_file );
+	if( ! under_lxde )
+		unlink( tmp_rc_file );
+
     return 0;
 }
